Put menu in a Gtk.Revealer
There is no more status area at the bottom of the window; the menu pops
up and covers the source view when needed.  It will take a little work
to make sure this never gets in the user’s way.

We have temporarily lost the functionality of the status area: showing
cursor position and save progress.